+++ weblate/trans/discovery.py
@@ -15,6 +15,9 @@ from django.utils.translation import gettext
 
 from weblate.formats.models import FILE_FORMATS
 from weblate.logger import LOGGER
+from weblate.trans.component_copy import (
+    get_inherited_component_fields,
+)
 from weblate.trans.defines import COMPONENT_NAME_LENGTH
 from weblate.trans.models import Component
 from weblate.trans.tasks import create_component
@@ -29,38 +32,7 @@ if TYPE_CHECKING:
 
     from weblate.formats.base import TranslationFormat
 
-# Attributes to copy from main component
-COPY_ATTRIBUTES = (
-    "project",
-    "vcs",
-    "license",
-    "agreement",
-    "source_language",
-    "report_source_bugs",
-    "hide_glossary_matches",
-    "allow_translation_propagation",
-    "contribute_project_tm",
-    "enable_suggestions",
-    "suggestion_voting",
-    "suggestion_autoaccept",
-    "check_flags",
-    "new_lang",
-    "language_code_style",
-    "commit_message",
-    "add_message",
-    "delete_message",
-    "merge_message",
-    "addon_message",
-    "pull_message",
-    "push_on_commit",
-    "commit_pending_age",
-    "edit_template",
-    "manage_units",
-    "variant_regex",
-    "category_id",
-    "key_filter",
-    "secondary_language",
-)
+COPY_ATTRIBUTES = get_inherited_component_fields("project", "category_id")
 
 
 class DiscoveryErrorMatch(TypedDict):
@@ -346,9 +318,9 @@ class ComponentDiscovery:
         for key in COPY_ATTRIBUTES:
             if key not in kwargs and main is not None:
                 kwargs[key] = getattr(main, key)
-        # Copy file format parameters if the format is same
-        if main is not None and self.file_format == main.file_format:
-            kwargs["file_format_params"] = main.file_format_params
+        if main is not None and self.file_format != main.file_format:
+            kwargs.pop("enforced_checks", None)
+            kwargs.pop("file_format_params", None)
 
         # Disable template editing if not supported by format
         if not self.file_format_cls.can_edit_base: